[fix] Rename viewer control components to prevent unplug warnings
Renamed Load3D viewer control components to use ViewerXxxControls naming convention
to avoid conflicts with the unplug system's component name detection. This prevents
false positive warnings when components are registered in the global scope.
