+++ src/components/load3d/Load3dViewerContent.vue
@@ -70,11 +70,11 @@
 import Button from 'primevue/button'
 import { onBeforeUnmount, onMounted, ref, toRaw } from 'vue'
 
-import CameraControls from '@/components/load3d/controls/viewer/CameraControls.vue'
-import ExportControls from '@/components/load3d/controls/viewer/ExportControls.vue'
-import LightControls from '@/components/load3d/controls/viewer/LightControls.vue'
-import ModelControls from '@/components/load3d/controls/viewer/ModelControls.vue'
-import SceneControls from '@/components/load3d/controls/viewer/SceneControls.vue'
+import CameraControls from '@/components/load3d/controls/viewer/ViewerCameraControls.vue'
+import ExportControls from '@/components/load3d/controls/viewer/ViewerExportControls.vue'
+import LightControls from '@/components/load3d/controls/viewer/ViewerLightControls.vue'
+import ModelControls from '@/components/load3d/controls/viewer/ViewerModelControls.vue'
+import SceneControls from '@/components/load3d/controls/viewer/ViewerSceneControls.vue'
 import { t } from '@/i18n'
 import { LGraphNode } from '@/lib/litegraph/src/LGraphNode'
 import { useLoad3dService } from '@/services/load3dService'
