+++ src/components/load3d/controls/viewer/ViewerLightControls.vue
